Removed the include_str!() stuff and use just the dummy book for testing
