+++ tests/dummy_book/src/SUMMARY.md
@@ -4,6 +4,7 @@
 
 - [First Chapter](./first/index.md)
     - [Nested Chapter](./first/nested.md)
+---
 - [Second Chapter](./second.md)
 
 [Conclusion](./conclusion.md)
